Move constructed inv tests to another file
